pyproject.toml
[project]
name = "mts-spi-study-cluster"
version = "0.1.0"
description = "PBS-ready driver for multivariate time-series SPI studies"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "pandas>=2.1",
    "pyyaml>=6.0",
    "pyspi>=1.1.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "tqdm>=4.66",
    "psutil>=5.9",
    "pyarrow>=15.0",
    "JPype1>=1.5",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
